Fix not shown output elments upon tab change
The default behaviour of shiny is not to show output elements
(.shiny-bound-output) on hidden parts. These lements are only shown when
the respective part is shown. This is implemented by listening to the
`shown` event, which needs to be triggered to inform Shiny that an
element is now visible.

The tab functionality is implemented using jquery ui. Since this
libarary is of course not aware of the requirements of Shiny we have to
trigger a `shown` event manually in case we change a tab. This allows
Shiny to show output elements when needed.

This closes #28.