Refactor bulmaTabs
`bulmaTabs` currently needs a `label` argument for creating the links in
the tab header. This approach is not only a unnecessary duplication, but
also can lead to duplicated `id` tags, since the id is created from the
label.

In this commit, we store the `label` as attribute (`data-label`) in the
tab content div and drop the `id` setting. In `bulmaTabs` we loop
through all elements passed by `...`, retrieve the `data-label`
attribute and assign a "unique" name as `id` and use this as anchor for
the `<a>` tag. The unique name is generated by drawing a random number
between 1 and 10000. This approach is taken from `tabsetPanel`.

This commit contains breaking changes, as the parameter `labels` is
dropped from `bulmaTabs`.